Rework server
