+++ internal/server/assets/fs.go
@@ -1,0 +1,6 @@
+package assets
+
+import "embed"
+
+//go:embed *
+var FS embed.FS
